+++ Cargo.toml
@@ -1,9 +1,10 @@
 [workspace]
 members = [
     "crates/kftray-server",
-    "crates/kftray-utils",
     "crates/kftray-tauri",
 	"crates/kftray-tui",
+	"crates/kftray-portforward",
+    "hacks/kftray-utils",
 ]
 
 resolver = "2"
