+++ crates/kftray-commons/.gitignore
@@ -1,0 +1,5 @@
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+
+
