+++ crates/kftray-commons/Cargo.toml
@@ -1,0 +1,42 @@
+[package]
+name = "kftray-commons"
+version = "0.12.2"
+description = "KFtray commons"
+authors = [
+    "Henrique Cavarsan <hencavarsan@gmail.com>",
+]
+license = "MIT"
+homepage = "https://kftray.app"
+repository = "https://github.com/hcavarsan/kftray"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.86"
+bytes = "1.6.0"
+dashmap = "6.0.0"
+flate2 = "1.0"
+httparse = "1.9.4"
+k8s-openapi = { version = "0.22.0", default-features = false, features = ["latest"] }
+lazy_static = "1.5.0"
+log = "0.4"
+rand = "0.8.5"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.120"
+tokio = { version = "1.39.1", features = ["rt-multi-thread", "macros", "full"] }
+tracing = "0.1.40"
+uuid = { version = "1.10.0", features = ["v4"] }
+dirs = "5.0.1"
+tauri = { version = "1.6", default-features = false, features = [
+	"updater",
+	"api-all",
+	"macos-private-api",
+	"system-tray",
+	"icon-png",
+	"devtools",
+] }
+sqlx = { version = "0.8.0", features = ["sqlite", "runtime-tokio-native-tls"] }
+hostsfile = { git = "https://github.com/tonarino/innernet", branch = "main" }
+
+[lib]
+name = "kftray_commons"
+path = "src/lib.rs"
