+++ crates/kftray-commons/src/lib.rs
@@ -1,0 +1,6 @@
+// src/lib.rs
+pub mod models;
+pub mod utils;
+
+pub use models::*;
+pub use utils::*;
