+++ crates/kftray-commons/src/models/config_model.rs
