+++ crates/kftray-commons/src/models/config_state_model.rs
