+++ crates/kftray-commons/src/models/window.rs
