+++ crates/kftray-commons/src/utils/db.rs
