+++ crates/kftray-commons/src/utils/logging.rs
