+++ crates/kftray-commons/src/utils/validate_configs.rs
