+++ crates/kftray-portforward/Cargo.toml
@@ -1,0 +1,41 @@
+[package]
+name = "kftray-portforward"
+version = "0.12.2"
+description = "KFtray library with port forwarding logic"
+authors = [
+    "Henrique Cavarsan <hencavarsan@gmail.com>",
+]
+license = "MIT"
+homepage = "https://kftray.app"
+repository = "https://github.com/hcavarsan/kftray"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.86"
+async-trait = "0.1"
+bytes = "1.6.0"
+dashmap = "6.0.0"
+flate2 = "1.0"
+futures = "0.3.30"
+httparse = "1.9.4"
+k8s-openapi = { version = "0.22.0", default-features = false, features = ["latest"] }
+kube = { version = "0.93.1", features = ["client", "config", "rustls-tls", "ws"] }
+kube-runtime = "0.93.1"
+lazy_static = "1.5.0"
+log = "0.4"
+rand = "0.8.5"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.120"
+tokio = { version = "1.39.1", features = ["rt-multi-thread", "macros", "full"] }
+tokio-stream = { version = "0.1.15", features = ["net"] }
+tracing = "0.1.40"
+uuid = { version = "1.10.0", features = ["v4"] }
+whoami = "1.5.1"
+hostsfile = { git = "https://github.com/tonarino/innernet", branch = "main" }
+kftray-commons = { path = "../kftray-commons" }
+tower = "0.4.13"
+hyper-util = "0.1.7"
+
+[lib]
+name = "kftray_portforward"
+path = "src/lib.rs"
