+++ crates/kftray-portforward/src/lib.rs
@@ -1,0 +1,13 @@
+// src/lib.rs
+pub mod client;
+pub mod core;
+pub mod models;
+pub mod pod_finder;
+pub mod port_forward;
+
+pub use core::*;
+
+pub use client::*;
+pub use models::*;
+pub use pod_finder::*;
+pub use port_forward::*;
