+++ crates/kftray-tauri/src/commands/mod.rs
@@ -1,0 +1,14 @@
+// Module: github
+pub mod github;
+// Module: portforward
+pub mod portforward;
+// Module: window_state
+pub mod window_state;
+// Module: kube
+pub mod kubecontext;
+// Module: httplogs
+pub mod httplogs;
+// Module: config
+pub mod config;
+// Module: config_state
+pub mod config_state;
