+++ crates/kftray-tauri/src/lib.rs
@@ -1,27 +1,5 @@
-// src/lib.rs
-
 pub mod commands;
-pub mod config;
-pub mod config_state;
-pub mod db;
-pub mod keychain;
-pub mod kubeforward;
-pub mod logging;
-pub mod migration;
-pub mod models;
-pub mod remote_config;
-pub mod utils;
-pub mod window;
-
 pub use commands::*;
-pub use config::*;
-pub use config_state::*;
-pub use db::*;
-pub use keychain::*;
-pub use kubeforward::*;
-pub use logging::*;
-pub use migration::*;
-pub use models::*;
-pub use remote_config::*;
-pub use utils::*;
-pub use window::*;
+pub mod logging;
+pub mod tray;
+pub mod window;
