--- crates/kftray-tauri/src/models/mod.rs
@@ -1,5 +1,0 @@
-pub mod config;
-pub mod config_state;
-pub mod kube;
-pub mod response;
-pub mod window;
