--- crates/kftray-tauri/src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod config_dir;
-pub mod validate_configs;
