+++ crates/kftray-tui/Cargo.toml
@@ -11,7 +11,9 @@ repository = "https://github.com/hcavarsan/kftray"
 edition = "2021"
 
 [dependencies]
-kftray-tauri = { path = "../kftray-tauri" }  # Adjust the path as necessary
+kftray-tauri = { path = "../kftray-tauri" }
+kftray-commons = { path = "../kftray-commons" }
+kftray-portforward = { path = "../kftray-portforward" }
 ratatui = "0.27"
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
