+++ hacks/kftray-utils/src/bump_version.rs
