+++ hacks/kftray-utils/src/generate_icons.rs
